Refactor import statements to use file_handlers module for CifFile, ConfidenceJsonFile, and NpzFile
